api-flask/requirements.txt
flask
openai
pydantic
numpy
opencv-python
Pillow
flask-cors